requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
altair==4.0.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.13.5
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
black==22.3.0
bleach==6.2.0
blinker==1.9.0
boto3==1.24.87
botocore==1.27.96
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.6.10
cryptography==44.0.0
cycler==0.12.1
dask==2022.9.2
databricks==0.2
databricks-cli==0.17.3
databricks-sql-connector==2.1.0
datasets==2.5.2
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.85.0
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
Flask==2.3.3
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.4.0
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.4.1
greenlet==3.1.1
grpcio==1.69.0
h11==0.12.0
h5py==3.12.1
httpcore==0.15.0
httpx==0.25.1
huggingface-hub==0.27.1
idna==3.10
importlib-metadata==5.2.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.11.0
kiwisolver==1.4.8
lazy-object-proxy==1.10.0
libclang==18.1.1
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.6.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.0
mlflow==2.0.1
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.13
mypy-extensions==1.0.0
narwhals==1.21.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.23.3
oauthlib==3.2.2
onnx==1.12.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opt_einsum==3.4.0
orjson==3.10.13
overrides==7.7.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.16.0
psutil==6.1.1
pure_eval==0.2.3
py==1.11.0
pyarrow==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==1.10.20
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pylint==2.15.3
Pympler==1.1
PyNaCl==1.5.0
pyparsing==3.2.1
pytest==7.1.3
pytest-cov==4.0.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2022.7.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.2
querystring-parser==1.2.4
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.6.2
scikit-learn==1.1.2
scipy==1.13.1
seaborn==0.12.0
segtok==1.5.11
semver==3.0.2
Send2Trash==1.8.3
sentencepiece==0.1.97
shap==0.46.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.20.4
streamlit==1.13.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
thrift==0.13.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.22.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.18.3
validators==0.34.0
waitress==2.1.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
xxhash==3.5.0
yake==0.4.8
yarl==1.18.3
zipp==3.21.0
